rename main module to memory
