+++ openhands/memory/__init__.py
@@ -1,4 +1,4 @@
 from openhands.memory.condenser import Condenser
-from openhands.memory.memory import LongTermMemory
+from openhands.memory.long_term_memory import LongTermMemory
 
 __all__ = ['LongTermMemory', 'Condenser']
